+++ option.go
@@ -51,18 +51,3 @@ type (
 	Option  func(*options)
 	options Config
 )
-
-// ContinueOnError allows watcher continues watching configuration even Config.Load fails to load the loader.
-func ContinueOnError() LoadOption {
-	return func(options *loadOptions) {
-		options.continueOnError = true
-	}
-}
-
-type (
-	// LoadOption configures Config.Load with specific options.
-	LoadOption  func(*loadOptions)
-	loadOptions struct {
-		continueOnError bool
-	}
-)