+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Removed
 - Remove konf.Default() to disallow loading configuration into the default Config (#180).
 - Remove ExplainOption from Config.Explain for always blurring sensitive information (#180).
+- Remove LoadOption from Config.Load (#184).
 
 ## [0.6.3] - 2024-02-23